fix: :fire: clean/simplify apps related code/parts
