+++ src/lib/reducers/index.js
@@ -19,7 +19,6 @@ export const getContent = proxy('content', content.getContent)
 export const getLocale = proxy('locale', locale.getLocale)
 export const getApps = proxy('apps', apps.getApps)
 export const isAppListFetching = proxy('apps', apps.isAppListFetching)
-export const isAppListForbidden = proxy('apps', apps.isAppListForbidden)
 export const getCurrentApp = proxy('apps', apps.getCurrentApp)
 export const hasFetched = proxy('apps', apps.hasFetched)
